% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_studies}
\alias{cancer_studies}
\title{Sample of Industry Cancer Trials from 2021}
\description{
Cancer clinical trials based on
a query where:
}
\details{
1. `study_type` is "Interventional".
2. `sponsor_type` is "Industry".
3. `date_range` is trials from 2021-01-01 or newer.
4. The `description` includes the keyword "cancer".
5. `phase` is reported (not NA).
6. `primary_purpose` is "Treatment".
7. `minimum_enrollment` is 100.
}
\keyword{data}
