% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condition_lookup}
\alias{condition_lookup}
\title{Conditions Lookup Table}
\description{
Data frame giving a normalisation mapping for the conditions field in the
clinical trials database
}
\keyword{data}
