% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_create_duckdb}
\alias{ctgov_create_duckdb}
\title{Create DuckDB Connection Object}
\usage{
ctgov_create_duckdb(basedir, dbdir = "ctgov_db_all", verbose = TRUE)
}
\arguments{
\item{basedir}{character giving the location that the flat-file pipe files
have been unziped}

\item{dbdir}{Location for database files. Should be a path to an existing
directory in the file system.}

\item{verbose}{logical flag; should progress messages be printed?;
defaults to \code{TRUE}}
}
\value{
a database connection object
}
\description{
This function creates a local DuckDB version of the full CTrialsGov database
from the pipe-deliminated flat files. The resulting connection returned by
the function can be queried directly or used with \code{ctgov_create_data}
to create a more de-normalized version for use with the other functions
contained in this package.
}
\details{
The function requires downloading and unzipping the current database dump
files found at \url{https://aact.ctti-clinicaltrials.org/pipe_files}. Given
their large size (around 1.4GB as of June 2022), we find it preferrable to
download the file directly through a browser or other command line tool
rather than through the R native functions, which are not well-suited to
to restarting a partial download.
}
\author{
Taylor B. Arnold, \email{taylor@dvlab.io}
}
