% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_disconnect}
\alias{ctgov_disconnect}
\title{Disconnect from the Database}
\usage{
ctgov_disconnect()
}
\value{
does not return any value; used only for side effects
}
\description{
This function disconnects all current database connections. This is 
done by removing the ctrialsgov:::.volatiles package environment. Users
can then reconnect either from the current process or other processes.
}
