% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ctgov_gantt_labeller}
\alias{ctgov_gantt_labeller}
\title{Create a Gantt Labeler for Timeline Tooltips}
\usage{
ctgov_gantt_labeller(x)
}
\arguments{
\item{x}{the data.frame object returned from a query.}
}
\value{
a string that can be used as a label in ggplotly
}
\description{
Create a Gantt Labeler for Timeline Tooltips
}
