% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ctgov_get_query_terms}
\alias{ctgov_get_query_terms}
\title{Query terms for the ClinicalTrials.gov dataset}
\usage{
ctgov_get_query_terms()
}
\value{
a named list of allowed categorical values for the query
}
\description{
Returns a list showing the available category levels for querying the data
with the \code{ctgov_query} function.
}
