% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_load_duckdb_file}
\alias{ctgov_load_duckdb_file}
\title{Load Database from DuckDB File}
\usage{
ctgov_load_duckdb_file(dbdir = NULL)
}
\arguments{
\item{dbdir}{a character string naming the location of the DuckDB file}
}
\value{
does not return any value; used only for side effects
}
\description{
Loads a version of the current active database from an existing DuckDB file.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
