% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_load_rds_file}
\alias{ctgov_load_rds_file}
\title{Load Database from RDS File}
\usage{
ctgov_load_rds_file(file, dbdir = NULL)
}
\arguments{
\item{file}{a character string naming a file; should have
the extension 'rds'}

\item{dbdir}{Location for the output database files. Should be a path to
an existing directory in the file system. By default will
place files in the location where the package is installed.}
}
\value{
does not return any value; used only for side effects
}
\description{
Loads a version of the current active database as a binary R
file.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
