% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ctgov_query_api}
\alias{ctgov_query_api}
\title{Wrapper function for the API to return all tables}
\usage{
ctgov_query_api(max_rows = 100L, ...)
}
\arguments{
\item{max_rows}{integer. The maximum number of rows to return;
Set to \code{NULL} to return all rows.}

\item{...}{Options passed to \code{ctgov_query}.}
}
\value{
a named list of tables
}
\description{
Passes arguments to the \code{ctgov_query} function, but returns the result
as a list that joins all of the matching data from the other tables. Note
that a few options have different defaults for the API.
}
