% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ctgov_query_endpoint}
\alias{ctgov_query_endpoint}
\title{Query endpoint table from the ClinicalTrials.gov dataset}
\usage{
ctgov_query_endpoint(data = NULL, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{data}{a dataset to search over; set to \code{NULL}
to use the full dataset that is currently
loaded}

\item{ignore_case}{logical. Should the search ignore
capitalization. The default is \code{TRUE}.}

\item{match_all}{logical. Should the results required matching
all the keywords? The default is \code{FALSE}.}
}
\value{
a tibble object queried from the loaded database
}
\description{
This function selects a subset of the clinical trials data by using a
a variety of different search parameters. These include free text search
keywords, range queries for the continuous variables, and exact matches for
categorical fields. The function \code{ctgov_get_query_terms} shows the
categorical levels for the latter. The function will either take the entire
dataset loaded into the package environment or a previously queried input.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
