% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_save_file}
\alias{ctgov_save_file}
\title{Save Current Database}
\usage{
ctgov_save_file(file)
}
\arguments{
\item{file}{a character string naming a file; should have
the extension 'rds'}
}
\value{
does not return any value; used only for side effects
}
\description{
Saves a version of the current active database as a binary R
file.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
