% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{ctgov_schema}
\alias{ctgov_schema}
\alias{ctgov_get_schema}
\alias{ctgov_set_schema}
\title{Get and Set the Default Schema}
\usage{
ctgov_schema(schema = NULL)
}
\arguments{
\item{schema}{the name of the schema. (Default is NULL - None)}
}
\description{
This function sets the schema in which tables in which
the CT Trials tables reside. 

Get the current schema eiter of the following.

ctgov_schema()
ctgov_get_schema()

Set the current schema with the following.

ctgov_schema(<SCHEMA NAME>)
ctgov_set_schema(<SCHEMA NAME>)

A return of "" from the get functions indicates a schema is not specified.
}
