% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ctgov_table_endpoint}
\alias{ctgov_table_endpoint}
\title{Return Full Table of Endpoint results}
\usage{
ctgov_table_endpoint()
}
\value{
a tibble object of the results
}
\description{
This function returns the currently loaded version of a dataset showing the
whether the reported data suggests that the study endpoint was either met or
not met. This is an algorithmically generated field and is likey to contain
errors.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
