% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ctgov_table_joined}
\alias{ctgov_table_joined}
\title{Return Full Table of Trials}
\usage{
ctgov_table_joined()
}
\value{
a tibble object of the results
}
\description{
This function returns the currently loaded version of the joined data
from the clinical trials database. This is the same as running the query
function with no parameters, but included for consistency.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
