% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ctgov_table_outcomes}
\alias{ctgov_table_outcomes}
\title{Return Full Table of Outcomes}
\usage{
ctgov_table_outcomes()
}
\value{
a tibble object of the results
}
\description{
This function returns the currently loaded version of the outcomes data
from the clinical trials database.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
