% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ctgov_table_references}
\alias{ctgov_table_references}
\title{Return Full Table of References}
\usage{
ctgov_table_references()
}
\value{
a tibble object of the results
}
\description{
This function returns the currently loaded version of the references data
from the clinical trials database.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
