% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ctgov_tfidf}
\alias{ctgov_tfidf}
\title{TF-IDF Keywords}
\usage{
ctgov_tfidf(
  ...,
  max_terms = 10000,
  tolower = TRUE,
  nterms = 5L,
  min_df = 0,
  max_df = 1
)
}
\arguments{
\item{...}{one or more vectors of text to search; must all be the
same length}

\item{max_terms}{maximum number of terms to consider for keywords}

\item{tolower}{should keywords respect the case of the raw terms}

\item{nterms}{number of keyord terms to include}

\item{min_df}{minimum proportion of documents that a term should
be present in to be included in the keywords}

\item{max_df}{maximum proportion of documents that a term should
be present in to be included in the keywords}
}
\value{
a character vector of detected keywords
}
\description{
Takes one or more vectors of text and returns a vector of keywords.
}
