% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ctgov_to_plotly.ctgov_bar_plot}
\alias{ctgov_to_plotly.ctgov_bar_plot}
\title{importFrom plotly ggplotly layout}
\usage{
\method{ctgov_to_plotly}{ctgov_bar_plot}(p, ...)
}
\description{
importFrom plotly ggplotly layout
}
