% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ctgov_to_plotly}
\alias{ctgov_to_plotly}
\title{Convert a ctrialsgov Visualization to Plotly}
\usage{
ctgov_to_plotly(p)
}
\arguments{
\item{tooltip}{the tooltip to be shown. (Default is gantt_labeller(x))}
}
\description{
Convert a ctrialsgov Visualization to Plotly
}
