% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utils.R
\name{has_term}
\alias{has_term}
\title{Does a Term Appear in a Vector of Strings?}
\usage{
has_term(s, pattern, ignore_case = TRUE)
}
\arguments{
\item{s}{the vector of strings.}

\item{pattern}{the pattern to search for.}

\item{ignore_case}{should the case be ignored? Default TRUE}
}
\description{
Does a Term Appear in a Vector of Strings?
}
