% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_studies}
\alias{cancer_studies}
\title{Sample of Industry Cancer Trials from 2021}
\description{
Cancer clinical trials based on a query where:
`study_type` is "Interventional";
`sponsor_type` is "Industry";
`date_range` is trials from 2021-01-01 or newer;
The `description` includes the keyword "cancer";
`phase` is reported (not NA);
`primary_purpose` is "Treatment";
`minimum_enrollment` is 100.
}
\keyword{data}
