% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_create_data}
\alias{ctgov_create_data}
\title{Initialize the connection}
\usage{
ctgov_create_data(con, verbose = TRUE)
}
\arguments{
\item{con}{an DBI connection object to the database}

\item{verbose}{logical flag; should progress messages be printed?;
defaults to \code{TRUE}}
}
\value{
does not return any value; used only for side effects
}
\description{
This function must be run prior to other functions in the package. It
creates a parsed and cached version of the clinical trials dataset in
memory in R. This makes other function calls relatively efficient. other
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
