% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ctgov_kwic}
\alias{ctgov_kwic}
\title{Keywords in Context}
\usage{
ctgov_kwic(
  term,
  text,
  names = NULL,
  n = Inf,
  ignore_case = TRUE,
  use_color = FALSE,
  width = 20L,
  output = c("cat", "character", "data.frame")
)
}
\arguments{
\item{term}{search term as a string}

\item{text}{vector of text to search}

\item{n}{number of results to return; default is Inf}

\item{ignore_case}{should search ignore case? default is TRUE}

\item{width}{how many characters to show as context}
}
\value{
a character vector of the search results
}
\description{
Take a
}
