% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ctgov_load_sample}
\alias{ctgov_load_sample}
\title{Load sample dataset}
\usage{
ctgov_load_sample(cancer_studies = FALSE)
}
\arguments{
\item{cancer_studies}{logical; should we load a currated list of cancer
clinical trials}
}
\value{
does not return any value; used only for side effects
}
\description{
This function loads a sample dataset for testing and prototyping purposes.
after running, all of the functions in the package can then be used with
this sample data. It consists of a random sample of trials that was available
from ClinicalTrials.gov at the time of the package creation.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
