% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ctgov_plot_timeline}
\alias{ctgov_plot_timeline}
\title{Plot a Timeline for a Set of Clinical Trials}
\usage{
ctgov_plot_timeline(
  x,
  mapping = aes_string(nct_id = "nct_id", start = "start_date", end =
    "primary_completion_date", y = "nct_id", fill = "phase")
)
}
\arguments{
\item{x}{the data.frame object returned from a query.}

\item{mapping}{the aesthetic mapping.}
}
\description{
Plot a Timeline for a Set of Clinical Trials
}
\seealso{
ctgov_gantt_labeller
}
