% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ctgov_plot_timeline}
\alias{ctgov_plot_timeline}
\title{Plot a Timeline for a Set of Clinical Trials}
\usage{
ctgov_plot_timeline(
  x,
  start_date = "start_date",
  completion_date = "primary_completion_date",
  label_column = "nct_id",
  color = label_column,
  tooltip = ctgov_gantt_labeller(x)
)
}
\arguments{
\item{x}{the data.frame object returned from a query.}

\item{start_date}{the start date column name. (Default is "start_date")}

\item{completion_date}{the date the trial is set to be complete.
(Default "primary_completion_date").
(Default is "primary_completion_date")}

\item{label_column}{the column denoting the labels for the y-axis.
(Default is "nct_id")}

\item{color}{the column to be used for coloring. (Default is label_column)}

\item{tooltip}{the tooltips for each of trials.
(Default is `ctgov_gantt_labeller(x)`).}
}
\description{
Plot a Timeline for a Set of Clinical Trials
}
\seealso{
ctgov_gantt_labeller
}
