% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ctgov_to_plotly}
\alias{ctgov_to_plotly}
\title{Convert a ctrialsgov Visualization to Plotly}
\usage{
ctgov_to_plotly(p, ...)
}
\arguments{
\item{p}{the plot returned by `ctgov_plot_timeline()`.}

\item{...}{currently not used.}
}
\value{
a Plotly object
}
\description{
Convert a ctrialsgov Visualization to Plotly
}
